extract AkkadianWordAnalysis
