+++ src/dictionary/ui/display/WordTitle.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import Word from 'dictionary/domain/Word'
 import { Markdown } from 'common/Markdown'
 import { Col, Row } from 'react-bootstrap'
-import Ipa from 'akkadian/ui/ipa'
 import transcriptionsToPhoneticSegments from 'akkadian/application/phonetics/segments'
+import { Ipa } from 'akkadian/ui/akkadianWordAnalysis'
 
 function wordIpa(word: Word): JSX.Element {
   try {
